Move to correct component
